叶晓波

+++ HotelSystem_client/src/blservice/Hotel_blservice.java
@@ -20,7 +20,7 @@ public interface Hotel_blservice {
 
 	public ArrayList<HotelInfoVO> getHotelFromName(String text);
 
-	public ArrayList<HotelInfoVO> getHotelFromGrade(ArrayList<HotelInfoVO> list, double grade);
+	public ArrayList<HotelInfoVO> getHotelFromGrade(ArrayList<HotelInfoVO> list, int grade);
 
 	public ArrayList<HotelInfoVO> getHotelFromPrice(ArrayList<HotelInfoVO> list, int minPrice, int maxPrice);
